Rename the FtpConfigProperties and add load properties feature
